backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
python-decouple==3.6
sqlparse==0.2.4
tzdata==2022.6
PyYAML==6.0
attrs==22.2.0
drf-spectacular==0.25.1
inflection==0.5.1 
jsonschema==4.17.3 
pyrsistent==0.19.3 
pytz==2022.7.1
uritemplate==4.1.1
djangorestframework-simplejwt==5.2.2 
pyjwt==2.6.0